feat: update tenant account deletion rules migration and related documentation
